Move SecureProxy to server
