+++ java/proto/server/BUILD
@@ -1,0 +1,41 @@
+#
+# Copyright 2023 The Project Oak Authors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+load("@io_grpc_grpc_java//:java_grpc_library.bzl", "java_grpc_library")
+load("@rules_java//java:defs.bzl", "java_proto_library")
+load("@rules_proto//proto:defs.bzl", "proto_library")
+
+package(
+    default_visibility = ["//visibility:public"],
+    licenses = ["notice"],
+)
+
+proto_library(
+    name = "secure_proxy_proto",
+    srcs = ["secure_proxy.proto"],
+    deps = ["//oak_remote_attestation/proto/v1:service_streaming_proto"],
+)
+
+java_proto_library(
+    name = "secure_proxy_java_proto",
+    deps = [":secure_proxy_proto"],
+)
+
+java_grpc_library(
+    name = "secure_proxy_java_grpc",
+    srcs = [":secure_proxy_proto"],
+    deps = [":secure_proxy_java_proto"],
+)
