+++ java/src/main/java/com/google/oak/server/BUILD
@@ -25,6 +25,8 @@ java_library(
     name = "server",
     srcs = glob(["*.java"]),
     deps = [
+        "//java/proto/server:secure_proxy_java_grpc",
+        "//java/proto/server:secure_proxy_java_proto",
         "//java/src/main/java/com/google/oak/crypto:encryptor",
         "//java/src/main/java/com/google/oak/crypto:server_encryptor",
         "//java/src/main/java/com/google/oak/crypto/hpke",
